# replace cargo.rs with cargo