--- a/.gitignore
+++ b/.gitignore
@@ -6,3 +6,4 @@ src/scopes
 target/
 *.py[cod]
 .vscode/**
+test_project/**/Cargo.lock