--- a/test_project/src/lib.rs
+++ b/test_project/src/lib.rs
@@ -1,6 +1,5 @@
 #[cfg(test)]
 mod tests {
     #[test]
-    fn it_works() {
-    }
+    fn it_works() {}
 }