--- a/test_project/test-crate3/src/main.rs
+++ b/test_project/test-crate3/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("Hello, world!");
+}