… Cargo.lock
… Cargo.toml
… src/racer/cargo.rs
… src/racer/nameres.rs
… test_project/Cargo.toml
… test_project/test-crate2/Cargo.toml
… test_project/test-crate2/src/lib.rs
… test_project/test-crate3/Cargo.toml
… test_project/test_fixtures/Cargo.toml
… test_project/test_fixtures/src/foo.rs
… test_project/test_fixtures/src/lib.rs
… test_project/test_fixtures/src/submod/bar.rs
… tests/system.rs
… tests/utils.rs