Current racer's file search has many problems.
Sometimes it fails to find dependency(see comments in #551), and it causes too much recursion(#831).
To resolve these problems, I replaced `cargo.rs` with cargo crate.
For clients which does sync io, this change can cause 2~3s delay when downloading dependencies, but other than that, it works well.

To work with cargo, current tests implementation have some problems.
- Currently we use `env::set_current_dir` for test_project, but it doesn't work with cargo. 
- It uses our own `TmpDir` implementation, which creates directory under `racer/`. So if a test panics it leaves tempdir in our normal file system. It's a bit annoying.

So I implemented new `with_test_project` function which copies whole files into tempdir and then run tests in that directory. And I replace racer's TmpDir with `tempfile` crate and remove `sync!` macro in each tests. Each test directories have unique names so there's no need to lock threads. Now tests are much faster.

In addition, I want to leave some notes about dependency caching.
[Here](https://github.com/kngwyu/racer-nightly/blob/cdf686b1a095cff7efc9583bfec1698e09dd1bdf/src/racer/core.rs#L766) we have cached dependencies for each manifest files(i.e. Cargo.toml), because racer calls `get_crate_file` many times in a session.
Compared to @birkenfeld's way in #838, it has 2 differences.

1. It doesn't have [caches of crate roots](https://github.com/kngwyu/racer-nightly/blob/93cea32faa36f85173a1c700234649962b099065/src/racer/core.rs#L766). It call's `cargo::util::important_paths::find_root_manifest_for_wd` everytime. This is because I think it can't be bottleneck.

2. It has dependency cache in `Session`, not in thread_local global variable. It can cause decrease in speed for clients using racer as library like `rls`. This is because I think to use `thread_local` make difficult to provide stateful APIs in the future, but I have no strong opinion about this.

And here's a result of benchmark, though it's not so general(We need more practical bench, but it will be a future task).
Before
```
test benches::bench_code_chunks     ... bench:      86,014 ns/iter (+/- 5,202)
test benches::bench_iter_stmts      ... bench:     206,687 ns/iter (+/- 35,788)
test benches::bench_mask_comments   ... bench:      30,087 ns/iter (+/- 4,644)
test benches::bench_mask_sub_scopes ... bench:     127,966 ns/iter (+/- 23,547)
```

After
```
test benches::bench_code_chunks     ... bench:      87,280 ns/iter (+/- 5,913)
test benches::bench_iter_stmts      ... bench:     222,251 ns/iter (+/- 22,692)
test benches::bench_mask_comments   ... bench:      30,376 ns/iter (+/- 3,099)
test benches::bench_mask_sub_scopes ... bench:     127,343 ns/iter (+/- 14,004)
```