@kngwyu 2018-05-19
Oh appveyor's TARGET=i686-pc-windows-gnu build fails with an add error.
@TedDriggs 
Could you please take a look at build log?