@birkenfeld 2018-05-22
@kngwyu I was wondering about doing this while overhauling cargo.rs - but I thought that the disruption of updating metadata might be undesired for racer.

If it is deemed acceptable, I'm all for this, a more detailed review will follow.