@TedDriggs 2018-05-22
@kngwyu I no longer have access to a local Windows box. The error suggests that it was looking for `make` - I don't recall that being present in the past.