@kngwyu 2018-05-23
@TedDriggs 
Thanks. I'll try to build again in my account.
@birkenfeld 
>  but I thought that the disruption of updating metadata might be undesired for racer.

Very good point. My implementation allow `cargo` to write `Cargo.lock` file. It can be annoying for some clients like `RLS`.
So I decided to modify implementation not to write lock file, by calling [resolve_with_previous](https://github.com/rust-lang/cargo/blob/master/src/cargo/ops/resolve.rs#L136) directly.
Please do not merge this PR until I finished change.