@kngwyu 2018-05-23
Oops, my comment was incorrect. This implementation don't write lock files, but sometimes write to registry, when downloading files.
So it can cause operation delay by locking registry file, if we use it simultaneously with other `cargo` operation.
@nrc 
Is this can be problematic for RLS?

We can also get dependencies using only lock file by [load_pkg_lockfile](https://github.com/rust-lang/cargo/blob/master/src/cargo/ops/lockfile.rs#L11), but by this way, we can't resolve patch.crates-io.