@kngwyu 2018-05-23
I think there's 3 ways to use `cargo` for dependency resolving.
### use  [resolve_ws_precisely](https://github.com/rust-lang/cargo/blob/master/src/cargo/ops/resolve.rs#L25)
Current implementation
- **pros**
  - we don't have to build our package to use racer
- **cons**
  - sometime racer downloads dependencies from crates.io
  - racer writes `~/.cargo/registory` and `Cargo.lock`

### use [resolve_with_previous](https://github.com/rust-lang/cargo/blob/master/src/cargo/ops/resolve.rs#L136)
- **pros**
  - same as above + racer don't write `Cargo.lock` file
- **cons**
  - sometime racer downloads dependencies from crates.io and writes `~/.cargo/registory` 
  - implementation is a bit complex

### use [load_pkg_lockfile](https://github.com/rust-lang/cargo/blob/master/src/cargo/ops/lockfile.rs#L11)
- **pros**
  - racer write nothing to our file system
- **cons**
  - We have to build our package before trying to complete

Personally, I prefer to first way, because it can reduce build time. And, I think for stand-alone clients to write `Cargo.lock` and registry is not problematic.
But, for rls, which call `cargo` internally, this can cause operation delay, because `cargo` waits until file lock is unlocked, though it will be at most 2~3s.
@nrc 
Do you think allow racer to write `Cargo.lock` or registry file can be problematic for rls?