@nrc 2018-05-24
It would be good to test to make sure. I think that modifying Cargo.lock is problematic because we have to watch that and changing it will cause us to recompile (and a full recompile, not an incremental one). I think that modifying the registry should be fine, because we don't watch that. However, if there are significant changes (thinking that the registry gets out of sync with the lock file), if that causes our builds to error out, it will be a problem (to explain a bit, we try and just call rustc, not Cargo, if Cargo.toml or .lock change we run Cargo and flush our caches, but we don't watch the registry, so if something changes there we might need to change the way we call rustc, but we don't realise that, possibly that could cause errrors).