@kngwyu 2018-05-24
@nrc 
Thanks for detailed explanation.
I'll try to modify codes not to write lock file and use it with RLS.