@kngwyu 2018-05-26
OMG our new shiny support for `patch.crates-io` doesn't work with RLS !!!
But other than that there seems to be no problem with RLS, though I'll debug a little more.