@matklad 2018-05-28
@kngwyu would it be possible to use cargo metadata sub command instead of linking to Cargo directly?