@kngwyu 2018-05-28
@matklad 
Maybe No.
Write to `Cargo.lock` causes full recompile to RLS so we want to avoid it.