@matklad 2018-05-28
Hm, if `cargo metadata` updates Cargo.lock, that means that the current lockfile is outdated anyway, so it has to be updated before the compilation anyway. 

>But, for rls, which calls cargo internally, this can cause operation delay, because cargo waits until file lock is unlocked, though it will be at most 2~3s.

Note that writing lockfile can really take unbounded time: if the project has a lot of git dependencies, Cargo needs to clone all those repositories, and that can take a lot of time. That said, you'll *need* to clone those dependencies to provide correct completions. 


I don't fully understand the constraints of racer and rls, but I imagine one way this could work is to have, inside racer, some kind of API "give me information about the project". Than, the stand-alone racer would use cargo metadata, and for racer inside rls, rls would provide this data using its own knowledge of the project (which it gets from Cargo). This way, standalone racer could avoid linking to Cargo, and for racer inside rls, we get the gurantee that rls is single source of truth.