@kngwyu 2018-05-28
@matklad 
> I imagine one way this could work is to have, inside racer, some kind of API "give me information about the project".

Hmm... 
Yeah it's better if we can avoid linking to cargo, but I'm concerned about overhead.
But, for RLS, indeed it's better to use its information to get src_paths.