@matklad 2018-05-28
>but I'm concerned about overhead.

Yeah, that is a valid concern. Parsing lockfile and making sure it is up-to-date is not instantaneous, so it's better to cache `cargo metadata` output, if possible. On the other hand, I think the overhead of spawning a process and parsing JSON shouldn't be that big, compared to the work you actually need to do.