@kngwyu 2018-05-29
Now I think to have different source of `src_path`s (RLS and `cargo metadata`), but I would like to set it as a future task, because I don't know much about RLS so I can't implement this kind of functionality immediately, and now we have many bugs to fix.
@matklad 
Thanks for your whole advice, it's really helpful.